\name{powerMRMA-package}
\alias{powerMRMA-package}
\alias{powerMRMA-pkg}
\docType{package}
\title{
\packageTitle{powerMRMA}
}
\description{
\packageDescription{powerMRMA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{powerMRMA}
\packageIndices{powerMRMA}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{powerMRMA}

Maintainer: \packageMaintainer{powerMRMA}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
